+++ src/cli.rs
@@ -30,7 +30,7 @@ use std::{collections::HashMap, path::PathBuf, str::FromStr};
 const OUTPUT_DATE: &str = "%Y-%m-%d";
 const OUTPUT_TIME: &str = "%H:%M:%S";
 
-#[derive(FromArgs, PartialEq, Debug)]
+#[derive(FromArgs, PartialEq, Eq, Debug)]
 /// Polkadot Staking Rewards CLI-App
 pub struct App {
 	#[argh(option, from_str_fn(date_from_string), short = 'f')]
@@ -39,7 +39,7 @@ pub struct App {
 	/// date to stop crawling for staking rewards. Format: "YYY-MM-DD HH:MM:SS"
 	#[argh(option, from_str_fn(date_from_string), short = 't')]
 	pub to: Option<NaiveDateTime>,
-	/// network to crawl for rewards. One of: [Polkadot, Kusama, Moonriver, MOVR, KSM, DOT]
+	/// network to crawl for rewards. One of: [Polkadot, Kusama, Moonriver, Moonbeam, Calamari, MOVR, GLMR, KSM, DOT, KMA]
 	#[argh(option, default = "Network::Polkadot", short = 'n')]
 	pub network: Network,
 	/// the fiat currency which should be used for prices
@@ -105,7 +105,7 @@ pub fn date_from_string(value: &str) -> Result<chrono::NaiveDateTime, String> {
 	Ok(time)
 }
 
-#[derive(PartialEq, Debug)]
+#[derive(PartialEq, Eq, Debug)]
 pub enum Network {
 	/// The Polkadot Network
 	Polkadot,
@@ -115,8 +115,10 @@ pub enum Network {
 	Moonriver,
 	/// The Moonbeam Network
 	Moonbeam,
-	/// The Astar network
+	/// The Astar Network
 	Astar,
+	/// The Calamari Network
+	Calamari,
 	/// Alepgh-zero solo-chain
 	Aleph,
 }
@@ -129,6 +131,7 @@ impl Network {
 			Self::Moonbeam => "moonbeam",
 			Self::Moonriver => "moonriver",
 			Self::Astar => "astar",
+			Self::Calamari => "calamari",
 			Self::Aleph => "aleph",
 		}
 	}
@@ -141,6 +144,7 @@ impl Network {
 			Self::Moonriver => 10u128.pow(18),
 			Self::Moonbeam => 10u128.pow(18),
 			Self::Astar => 10u128.pow(18),
+			Self::Calamari => 10u128.pow(12),
 			Self::Aleph => 10u128.pow(12),
 		};
 		let frac = FixedU128::checked_from_rational(*amount, denominator)
@@ -159,11 +163,9 @@ impl FromStr for Network {
 			"moonriver" | "movr" => Ok(Network::Moonriver),
 			"moonbeam" | "glmr" => Ok(Network::Moonbeam),
 			"astar" | "astr" => Ok(Network::Astar),
+			"calamari" | "kma" => Ok(Network::Calamari),
 			"aleph" | "aleph-zero" | "azero" => Ok(Network::Aleph),
-			_ => bail!(
-				"Network must be one of: 'kusama', 'polkadot', 'moonbeam', 'astar', 'moonriver', 'aleph-zoer', or their
-				token abbreviations."
-			),
+			_ => bail!( "Network must be one of: 'kusama', 'polkadot', 'moonbeam', 'astar', 'moonriver', 'aleph-zoer', or their token abbreviations."),
 		}
 	}
 }
@@ -267,7 +269,10 @@ fn construct_progress_bar() -> Result<ProgressBar, Error> {
 	let bar = ProgressBar::new(1000);
 	bar.set_style(
 		ProgressStyle::default_bar()
-			.template("{spinner:.blue} {msg} [{elapsed_precise}] [{bar:40.cyan/blue}] {percent}% ({eta})")?
+			.template("{spinner:.blue} {msg} [{elapsed_precise}] [{bar:40.cyan/blue}] {percent}% ({eta})")
+			.expect(
+				"Progress template is invalid, this should not happen if formatted correctly. Please file bug  report.",
+			)
 			.progress_chars("#>-"),
 	);
 	Ok(bar)